Bug fix: unify(a, a) caused a crash.
Previously, an attempt to unify two type variables that were already
synonymous would cause a crash in the following code:

    del self.__inferred_types[set_x]
    del self.__inferred_types[set_y]

since we can't delete the same key from self.__inferred_types twice.
Also, even if we fixed that, the following line:

    new_set = self.__inferred_types.union(set_x, set_y)

would have failed because union() requires that its arguments are
distinct sets.

We solve this by short-cutting the unification process if we see that
the arguments have already been unified.